+++ packages/openfort-react/src/components/Openfort/OpenfortProvider.tsx
@@ -345,7 +345,7 @@ export const OpenfortProvider = ({
   return createElement(
     Openfortcontext.Provider,
     { value },
-    <Web3ContextProvider enabled={open}>
+    <Web3ContextProvider>
       <CoreOpenfortProvider
         openfortConfig={{
           baseConfiguration: {