+++ packages/openfort-react/src/components/Openfort/types.ts
@@ -172,6 +172,8 @@ type OpenfortUIOptions = {
   linkWalletOnSignUp?: LinkWalletOnSignUpOption
 
   authProviders: UIAuthProvider[]
+  authProvidersLength?: number
+
   skipEmailVerification?: boolean
   termsOfServiceUrl?: string
   privacyPolicyUrl?: string