+++ packages/openfort-react/src/components/contexts/web3/index.tsx
@@ -1,19 +1,7 @@
-/**
- * Provides reusable Web3 state and helpers to descendant components.
- *
- * @example
- * ```tsx
- * <Web3ContextProvider>
- *   <WalletSelector />
- * </Web3ContextProvider>
- * ```
- */
-
-import { createContext, useContext } from 'react'
+import { createContext } from 'react'
 
 import type { Address, Chain } from 'viem'
 import { useAccount } from 'wagmi'
-import { useWalletConnectUri } from '../../../hooks/connectors/useWalletConnectUri'
 import { useChainIsSupported } from '../../../hooks/useChainIsSupported'
 import { useChains } from '../../../hooks/useChains'
 
@@ -41,21 +29,12 @@ const Web3Context = createContext({
   account: undefined,
 } as Web3Context)
 
-export const Web3ContextProvider = ({ enabled, children }: { enabled?: boolean; children: React.ReactNode }) => {
-  const { uri: walletConnectUri } = useWalletConnectUri({
-    enabled,
-  })
-
+export const Web3ContextProvider = ({ children }: { children: React.ReactNode }) => {
   const { address: currentAddress, chain } = useAccount()
   const chainIsSupported = useChainIsSupported(chain?.id)
   const chains = useChains()
 
   const value = {
-    connect: {
-      getUri: (_id?: string) => {
-        return walletConnectUri
-      },
-    },
     dapp: {
       chains,
     },
@@ -81,4 +60,4 @@ export const Web3ContextProvider = ({ enabled, children }: { enabled?: boolean; 
  * const { account } = useWeb3();
  * ```
  */
-export const useWeb3 = () => useContext(Web3Context)
+// export const useWeb3 = () => useContext(Web3Context)